Lv3 구현

+++ src/CalculatorHW/AbstractOperation.java
@@ -1,5 +1,5 @@
 package CalculatorHW;
 
 public abstract class AbstractOperation {
-    public abstract double operate(int firstNumber, int secondNumber);
+    public abstract double operate(double firstNumber, double secondNumber);
 }
